refactor: 인터페이스 대신 상속을 사용해 중복 제거
